shutils/chroot.sh: always set HOME (to /tmp).
It took me a while to figure it out why the firefox-0.25 build
failed to install. At install time it execs the 'xpcshell' binary
and it segfaults if $HOME is unset.

Clearly xpcshell must check that getenv("HOME") is not NULL and I'll
report the issue upstream.

Anyway also set HOME in here just to avoid such problems in the future.